Move EventService to Timescout.Application project
