+++ TimeScout.Application/Interfaces/IEventService.cs
@@ -1,7 +1,6 @@
-using System;
-using TimeScout.API.Models;
+using TimeScout.Domain.Entities;
 
-namespace TimeScout.API.Services;
+namespace TimeScout.Application.Interfaces;
 
 public interface IEventService
 {
