+++ TimeScout.Application/Services/EventService.cs
@@ -1,8 +1,8 @@
-using System;
-using TimeScout.API.Models;
-using TimeScout.API.Repository;
+using TimeScout.Domain.Entities;
+using TimeScout.Domain.Interfaces;
+using TimeScout.Application.Interfaces;
 
-namespace TimeScout.API.Services;
+namespace TimeScout.Application.Services;
 
 public class EventService : IEventService
 {
